Move role types to a separate package
