+++ pkg/services/org/model.go
@@ -2,9 +2,10 @@ package org
 
 import (
 	"errors"
-	"fmt"
 	"strings"
 	"time"
+
+	roletypes "github.com/grafana/grafana/pkg/models/role_types"
 )
 
 // Typed errors
@@ -37,8 +38,7 @@ type OrgUser struct {
 	Updated time.Time
 }
 
-// swagger:enum RoleType
-type RoleType string
+type RoleType = roletypes.RoleType
 
 const (
 	RoleViewer RoleType = "Viewer"
@@ -154,60 +154,6 @@ type RemoveOrgUserCommand struct {
 	OrgID                    int64
 	ShouldDeleteOrphanedUser bool
 	UserWasDeleted           bool
-}
-
-func (r RoleType) IsValid() bool {
-	return r == RoleViewer || r == RoleAdmin || r == RoleEditor
-}
-
-func (r RoleType) Includes(other RoleType) bool {
-	if r == RoleAdmin {
-		return true
-	}
-
-	if r == RoleEditor {
-		return other != RoleAdmin
-	}
-
-	return r == other
-}
-
-func (r RoleType) Children() []RoleType {
-	switch r {
-	case RoleAdmin:
-		return []RoleType{RoleEditor, RoleViewer}
-	case RoleEditor:
-		return []RoleType{RoleViewer}
-	default:
-		return nil
-	}
-}
-
-func (r RoleType) Parents() []RoleType {
-	switch r {
-	case RoleEditor:
-		return []RoleType{RoleAdmin}
-	case RoleViewer:
-		return []RoleType{RoleEditor, RoleAdmin}
-	default:
-		return nil
-	}
-}
-
-func (r *RoleType) UnmarshalText(data []byte) error {
-	// make sure "viewer" and "Viewer" are both correct
-	str := strings.Title(string(data))
-
-	*r = RoleType(str)
-	if !r.IsValid() {
-		if (*r) != "" {
-			return fmt.Errorf("invalid role value: %s", *r)
-		}
-
-		*r = RoleViewer
-	}
-
-	return nil
 }
 
 type ByOrgName []*UserOrgDTO
